separate flatgeobuf crate
