+++ Cargo.toml
@@ -1,4 +1,15 @@
 [workspace]
-members = ["rust/geoarrow"]
+members = ["rust/geoarrow", "rust/geoarrow-flatgeobuf"]
 exclude = ["js"]
 resolver = "2"
+
+[workspace.package]
+# Package version for rust geoarrow modules
+version = "0.4.0-beta.1"
+authors = ["Kyle Barron <kylebarron2@gmail.com>"]
+edition = "2021"
+homepage = "https://geoarrow.org/geoarrow-rs/"
+repository = "https://github.com/geoarrow/geoarrow-rs"
+license = "MIT OR Apache-2.0"
+categories = ["science::geo"]
+rust-version = "1.82"
