+++ rust/geoarrow-flatgeobuf/Cargo.toml
@@ -1,67 +1,53 @@
 
 [package]
-name = "geoarrow"
-version = "0.4.0-beta.1"
-authors = ["Kyle Barron <kylebarron2@gmail.com>"]
-edition = "2021"
-license = "MIT OR Apache-2.0"
-repository = "https://github.com/geoarrow/geoarrow-rs"
+name = "geoarrow-flatgeobuf"
+version = { workspace = true }
+authors = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
 description = "Rust implementation of GeoArrow"
-categories = ["science::geo"]
-rust-version = "1.82"
+categories = { workspace = true }
+rust-version = { workspace = true }
 
 [features]
-csv = ["dep:arrow-csv", "geozero/with-csv"]
-gdal = ["dep:gdal"]
-geos = ["dep:geos"]
-ipc_compression = ["arrow-ipc/lz4", "arrow-ipc/zstd"]
-parquet = ["dep:parquet"]
-parquet_async = [
-  "parquet",
-  "parquet/async",
-  "dep:async-stream",
-  "dep:futures",
-  "dep:tokio",
+default = ["async"]
+async = [
+  "flatgeobuf/http",
+  "dep:async-trait",
+  "dep:bytes",
+  "dep:http-range-client",
+  "dep:object_store",
 ]
-parquet_compression = [
-  "parquet/snap",
-  "parquet/brotli",
-  "parquet/flate2",
-  "parquet/lz4",
-  "parquet/zstd",
-]
-polylabel = ["dep:polylabel"]
-postgis = ["dep:futures", "dep:sqlx"]
-proj = ["dep:proj"]
-rayon = ["dep:rayon"]
-
 
 [dependencies]
-arrow = { version = "53.3", features = ["ffi"] }
-arrow-array = { version = "53.3", features = ["chrono-tz"] }
-arrow-buffer = "53.3"
 arrow-cast = { version = "53.3" }
 arrow-csv = { version = "53", optional = true }
 arrow-data = "53.3"
 arrow-ipc = "53.3"
 arrow-schema = "53.3"
-async-stream = { version = "0.3", optional = true }
+async-trait = { version = "0.1", optional = true }
+bytes = { version = "1.5.0", optional = true }
 chrono = { version = "0.4" }
 dbase = "0.5.0"
 enum-as-inner = "0.6.1"
 # Set default-features = false because async not working in wasm right now
 # For geo-traits impl
-flatgeobuf = { git = "https://github.com/flatgeobuf/flatgeobuf", rev = "f554f2768b612e131e9f55d014eaa5b911a7f1b5", optional = true, default-features = false }
+flatgeobuf = { git = "https://github.com/flatgeobuf/flatgeobuf", rev = "f554f2768b612e131e9f55d014eaa5b911a7f1b5", default-features = false }
 futures = { version = "0.3", optional = true }
 gdal = { version = "0.17", optional = true }
-geo = "0.28"
+geo = "0.29.3"
+geoarrow = { path = "../geoarrow" }
 geo-index = "0.1.1"
 geo-traits = "0.2"
 geos = { version = "9.0", features = ["v3_10_0", "geo"], optional = true }
 geozero = { version = "0.14", features = ["with-wkb"] }
 half = { version = "2.4.1" }
+http-range-client = { version = "0.8", optional = true }
 indexmap = { version = "2" }
 lexical-core = { version = "0.8.5" }
+object_store = { version = "0.11", optional = true }
 parquet = { version = "53", optional = true, default-features = false, features = [
   "arrow",
 ] }
@@ -107,40 +93,6 @@ parquet = { version = "53", default-features = false, features = [
 [lib]
 doctest = true
 
-[[example]]
-name = "gdal"
-test = false
-required-features = ["gdal"]
-
-[[bench]]
-name = "area"
-harness = false
-required-features = ["flatgeobuf"]
-
-[[bench]]
-name = "from_geo"
-harness = false
-
-[[bench]]
-name = "geos_buffer"
-harness = false
-required-features = ["geos"]
-
-[[bench]]
-name = "nybb"
-harness = false
-required-features = ["ipc_compression"]
-
-[[bench]]
-name = "translate"
-harness = false
-
-[[bench]]
-name = "wkb"
-harness = false
-bench = false                               # TODO fix this benchmark
-required-features = ["parquet_compression"]
-
 [package.metadata.docs.rs]
 rustdoc-args = ["--cfg", "docsrs"]
-features = ["csv", "flatgeobuf", "geos", "parquet", "postgis", "rayon"]
+features = ["async"]
