+++ rust/geoarrow-flatgeobuf/src/flatgeobuf/mod.rs
@@ -3,7 +3,7 @@
 mod reader;
 mod writer;
 
-#[cfg(feature = "flatgeobuf_async")]
+#[cfg(feature = "async")]
 pub use reader::read_flatgeobuf_async;
 pub use reader::{read_flatgeobuf, FlatGeobufReaderOptions};
 pub use writer::{write_flatgeobuf, write_flatgeobuf_with_options};
