+++ rust/geoarrow-flatgeobuf/src/flatgeobuf/reader/common.rs
