+++ rust/geoarrow-flatgeobuf/src/flatgeobuf/reader/mod.rs
@@ -1,11 +1,11 @@
-#[cfg(feature = "flatgeobuf_async")]
+#[cfg(feature = "async")]
 mod r#async;
 mod common;
-#[cfg(feature = "flatgeobuf_async")]
+#[cfg(feature = "async")]
 mod object_store_reader;
 mod sync;
 
 pub use common::FlatGeobufReaderOptions;
-#[cfg(feature = "flatgeobuf_async")]
+#[cfg(feature = "async")]
 pub use r#async::read_flatgeobuf_async;
 pub use sync::read_flatgeobuf;
