+++ rust/geoarrow-flatgeobuf/src/flatgeobuf/reader/object_store_reader.rs
