+++ rust/geoarrow-flatgeobuf/src/flatgeobuf/reader/sync.rs
