+++ rust/geoarrow-flatgeobuf/src/lib.rs
@@ -1,0 +1,3 @@
+#![cfg_attr(not(test), deny(unused_crate_dependencies))]
+
+mod flatgeobuf;
